Factorising xilinx vivado code
